+++ src/util/result2profile.cpp
@@ -293,7 +293,7 @@ int result2profile(DBReader<unsigned int> &resultReader, Parameters &par, const 
                 }
                 // write query, consensus sequence and neffM
                 result.push_back(static_cast<unsigned char>(centerSequence.int_sequence[pos]));
-                result.push_back(static_cast<unsigned char>(subMat.aa2int[pssmRes.consensus[pos]]));
+                result.push_back(static_cast<unsigned char>(subMat.aa2int[static_cast<int>(pssmRes.consensus[pos])]));
                 unsigned char neff = MathUtil::convertNeffToChar(pssmRes.neffM[pos]);
                 result.push_back(neff);
             }
